% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd_rlm}
\alias{fd_rlm}
\title{Fit Robust Triangular Fundamental Diagram}
\usage{
fd_rlm(flow, occupancy, cutoff = 0.15, ...)
}
\arguments{
\item{flow}{Count of passed vehicles}

\item{occupancy}{Traffic density between 0 and 1}

\item{cutoff}{Point separating high and low occupancy}

\item{...}{Additional arguments for code{\link[MASS]{rlm}}}
}
\value{
fd list containing fitted FD
}
\description{
The fundamental diagram (FD) in traffic engineering represents 
flow as a function of density.
}
\details{
This only fits 2 robust linear models and returns the coefficients. It
is not a true triangular fundamental diagram because the fitted models
include intercepts, and does not enforce the constraint that low and
high density should meet at the cutoff point.
}
