% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fun_diagram}
\alias{fun_diagram}
\title{Fit the fundamental diagram to observations}
\usage{
fun_diagram(flow, density, method = "ls", cutoff = 0.15)
}
\arguments{
\item{flow}{Count of passed vehicles}

\item{density}{Traffic density between 0 and 1}

\item{method}{Method for curve fitting. Available options: \code{"ls"} for least squares.}

\item{cutoff}{Point separating high and low density}
}
\value{
fd Object of class \code{fd}.
}
\description{
The fundamental diagram in traffic engineering represents 
flow as a function of density.
}
