% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pems.R
\name{read30sec}
\alias{read30sec}
\title{Read a 30 second raw PeMS file}
\usage{
read30sec(file, lanes = 1:2, numlanes = 8, nrows = 10000L, ...)
}
\arguments{
\item{file}{Name of file to read}

\item{lanes}{Vector of lanes to read. Others ignored.}

\item{numlanes}{Total expected number of lanes in file}

\item{nrows}{Number of rows to read. Use -1 to read all of them.}

\item{...}{Additional parameters for read.table}
}
\value{
data.frame
}
\description{
Takes a conservative approach to cleaning data by only performing the following steps:
Occupancy greater than 1 is converted to NA
}
\details{
ID is the station id. Name chosen to match metadata files.
}
