% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pems.R
\name{split_write}
\alias{split_write}
\title{Split And Append Results To CSV File}
\usage{
split_write(x, f, dirname, ...)
}
\arguments{
\item{x}{data frame to split}

\item{f}{factor defining splits}

\item{dirname}{character directory, will be created if doesn't exist}

\item{...}{further arguments to split}
}
\description{
x will be split by f and each group will be appended to a directory of
csv files
}
