% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pems.R
\name{read30sec}
\alias{read30sec}
\title{Read 30 Second Raw PeMS File}
\usage{
read30sec(file, posix_timestamp = FALSE, lanes = 1:2, numlanes = 8,
  nrows = 10000L, ...)
}
\arguments{
\item{file}{Name of file to read}

\item{posix_timestamp}{Convert the timestamp to POSIXct}

\item{lanes}{Vector of lanes to read. Others ignored.}

\item{numlanes}{Total expected number of lanes in file}

\item{nrows}{Number of rows to read. Use -1 to read all of them.}

\item{...}{Additional parameters for read.table}
}
\value{
data.frame
}
\description{
Read 30 Second Raw PeMS File
}
